Refactor JetpackRequestBuilder instead of using extension functions
